ml-service/ml_pipeline/requirements.txt
# ── Core Data Science ─────────────────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.0
imbalanced-learn>=0.12.0

# ── Visualization ─────────────────────────────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0

# ── MongoDB ───────────────────────────────────────────────────────────────────
pymongo>=4.7.0

# ── API (Optional) ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
joblib>=1.4.0
tabulate>=0.9.0
shap>=0.45.0
scipy>=1.12.0
mlflow>=2.12.0
